requirements.txt
Django==1.9.2
djangorestframework==3.3.2
celery==3.1.20
django-celery==3.1.17
redis==2.10.5
requests==2.9.1
requests-cache==0.4.10
psycopg2==2.6.1
PyYAML==3.11
GitPython==1.0.2
django-compressor==2.0
django-filter==0.12.0
libsass==0.10.1
django-dynamic-fixture==1.8.5
mock==1.3.0
-e git://git.linaro.org/lava/django-crowd-rest-backend.git@c741f1738786c58253b400ffc27a16967702f84c#egg=django_crowd_rest_backend-dev

pytz
pdbpp
ndg-httpsclient
cryptography
pyOpenSSL
